+++ backend/.env
@@ -1,4 +1,5 @@
-DATABASE_URL = "postgresql://postgres:1234@localhost/assetManagement"
+POSTGRESSQL_URL = "postgresql://postgres:1234@localhost/assetManagement"
+
 SESSION_KEY = 'sakfjlksdfkl393eif9jif3if9ji3fj'
 GOOGLE_CLIENT_ID = '1031994067890-ghnonjiptj3ans82fh7cd4mpf03m5nhs.apps.googleusercontent.com'
 GOOGLE_CLIENT_SECRET = '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'
