+++ backend/.gitignore
@@ -14,7 +14,8 @@ test.db
 log.txt
 Pipfile.lock
 env3.*
-env
+.env
+*.env*
 docs_build
 site_build
 venv
