+++ backend/api/v1/auth/service/auth_service.py
@@ -3,34 +3,12 @@ from sqlalchemy.orm import Session
 from starlette.requests import Request
 from fastapi import HTTPException
 from authlib.integrations.starlette_client import OAuthError
-from uuid import uuid4
-from datetime import datetime
 # Module
-from auth.database.models import User
-from auth.database.schemas import ProviderEnum, UserRoleEnum
-from auth.service.google_service import authenticate_with_google
-from auth.service.kakao_service import authenticate_with_kakao
-from auth.service.naver_service import authenticate_with_naver
-
-
-class DBHandler():
-    def get_user(self, db: Session, social_id: str, provider: ProviderEnum):
-        return db.query(User).filter(User.social_id == social_id, User.provider == provider.value).first()
-
-    def create_user(self, db: Session, social_id: str, provider: ProviderEnum, role: UserRoleEnum=None, nickname: str=None):
-        new_user = User(
-                id=uuid4(),  
-                social_id=social_id,
-                provider=provider,
-                role=role,
-                nickname=nickname,
-                created_at=datetime.now()  
-            )
-        
-        db.add(new_user)
-        db.commit()
-        db.refresh(new_user)  
-        return new_user
+from api.v1.auth.database.schemas import ProviderEnum
+from api.v1.auth.service.google_service import authenticate_with_google
+from api.v1.auth.service.kakao_service import authenticate_with_kakao
+from api.v1.auth.service.naver_service import authenticate_with_naver
+from api.v1.auth.database.repository import DBHandler
 
 class AuthenticationBuilder(DBHandler):
     async def naver_authenticate(self, db:Session, request: Request):
